feat: SDL3 port

+++ Minecraft.Client/PlayerConnection.cpp
@@ -1274,7 +1274,7 @@ void PlayerConnection::handleSetCreativeModeSlot(shared_ptr<SetCreativeModeSlotP
             // 4J-PB - for Xbox maps, we'll centre them on the origin of the world, since we can fit the whole world in our map
             data->x = centreXC;
             data->z = centreZC;
-            data->dimension = (byte)player->level->dimension->id;
+            data->dimension = player->level->dimension->id;
             data->setDirty();
         }
