+++ Minecraft.World/Inventory.cpp
@@ -536,7 +536,7 @@ ListTag<CompoundTag> *Inventory::save(ListTag<CompoundTag> *listTag)
         if (items[i] != NULL)
         {
             CompoundTag *tag = new CompoundTag();
-            tag->putByte(L"Slot", (byte)i);
+            tag->putByte(L"Slot", i);
             items[i]->save(tag);
             listTag->add(tag);
         }
@@ -546,7 +546,7 @@ ListTag<CompoundTag> *Inventory::save(ListTag<CompoundTag> *listTag)
         if (armor[i] != NULL)
         {
             CompoundTag *tag = new CompoundTag();
-            tag->putByte(L"Slot", (byte)(i + 100));
+            tag->putByte(L"Slot", (i + 100));
             armor[i]->save(tag);
             listTag->add(tag);
         }
