+++ Minecraft.World/Ozelot.cpp
@@ -60,7 +60,7 @@ void Ozelot::defineSynchedData()
 {
     TamableAnimal::defineSynchedData();
 
-    entityData->define(DATA_TYPE_ID, (byte)TYPE_OZELOT);
+    entityData->define(DATA_TYPE_ID, TYPE_OZELOT);
 }
 
 void Ozelot::serverAiMobStep()
@@ -309,7 +309,7 @@ int Ozelot::getCatType()
 
 void Ozelot::setCatType(int type)
 {
-    entityData->set(DATA_TYPE_ID, (byte)type);
+    entityData->set(DATA_TYPE_ID, type);
 }
 
 bool Ozelot::canSpawn()
